Big do_rain() Functionality Update
Rain has now complete audio functionality where the sound will fade depending on if there was a change in the rain type.

Also fixed minor bug with weather type change, forgot to add the time the weather started to the duration of the weather type.  Caused reinitialization of weather type every frame.