Weather Syncing
Weather should now be synced regardless of time tabbed out or when tabbed out because all times are now based on the gateDropTime seed and then the previous weather duration and time started.

It now acts like a chain so if someone tabs into a session 10 minutes in they will be at the same time throughout some weather cycle as someone who's been playing the entire time.