Complete Rework of Rain/Thunder
Completely reworked the way rain/thunder is handled.  So far I've changed the rain function to no longer check if the rain type is light, medium, or heavy every time it's needed. It will instead check it only once and set a current rain type and that current rain type will hold an object with all the necessary sounds, billboards, and volumes. That current rain type is used for all the work instead of all the extra checks.

Thunder and lightning now works on the same type of system that the weather cycles work, it stores each lightning strike in an array and goes through different cycles. It also now bases all the next lightning strike times off of a shared seed value for all clients so it won't matter if someone's second counter is slightly different, everyone will see them at the same time.  So the thunder & lightning system now works so that its both seeded and can be ran back in demos and be played at the exact same time every time.

I've been stuck trying to get the thunder/lightning system to account for tabbed out players or players that join late because the current weather could be completely different from the weather they tabbed out or joined during. I will have to try to find a work around.

I also added a no rain spots a while ago where if the camera is in a spot where rain should be hitting an object it will move the billboard to the specified billboard height in that area. It uses vertices to draw a polygon and checks if the camera is in that polygon.

I also added a smite system a long time ago as a joke, it will smite people with specified names in the list and then go down the list and start picking off riders 1 by 1 crashing their game as if they died.