Rain Billboard Rework
Reworked the rain billboard system to not have the need for any billboards in the billboards file. Instead it tries to find rain billboards and if there's not enough it will keep adding.  If there's excess it will hide the excess billboards.

- Each rain type now posseses 4 new properties:

- size: holds the size for each rain billboard
- aspect: holds the aspect ratio for each rain billboard
- maxBillboardHeight: the maximum height of the billboard in feet before the it starts following the camera
- billboardIndexMap: an array that maps a billboardArr index to a rain billboard index.

With these new features and properties the indexStart property of the rain objects became obsolete.